Move LoginComponent and HomeModule to _shared folder"
* _shared is currently used as an alias for _components
* Enabled absolute import paths. Currently only used in Login and Home.
