+++ src/app/_shared/home/home.component.ts
