+++ src/app/_shared/login/login.component.spec.ts
@@ -2,10 +2,10 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { LoginComponent } from './login.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { AuthService } from '../_services/auth/auth.service';
+import { AuthService } from '_services/auth/auth.service';
 import { HttpClient, HttpHandler } from '@angular/common/http';
-import { ConfigurationService } from '../_services';
-import { TokenStorage } from '../_classes';
+import { ConfigurationService } from '_services/configuration/configuration.service';
+import { TokenStorage } from '_classes/tokenstorage/tokenstorage';
 import { RouterTestingModule } from '@angular/router/testing';
 
 describe('LoginComponent', () => {
