+++ tsconfig.json
@@ -8,12 +8,15 @@
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "target": "es5",
-    "typeRoots": [
-      "node_modules/@types"
-    ],
-    "lib": [
-      "es2017",
-      "dom"
-    ]
+    "typeRoots": ["node_modules/@types"],
+    "lib": ["es2017", "dom"],
+    "baseUrl": "src",
+    "paths": {
+      "_classes/*": ["app/_classes/*"],
+      "_guards/*": ["app/_guards/*"],
+      "_shared/*": ["app/_shared/*"],
+      "_services/*": ["app/_services/*"],
+      "app/*": ["app/*"],
+    }
   }
 }
